removed displayTiddler hijack
This was previously added so that if a tiddler was discovered that
was created by the activity stream (e.g. via the excludeLists tag)
was opened it would update to be the server version. However this
ignores the fact that if you click edit, the same thing happens having
bad side effects.

Thus removing this code as it's not worth the hassle.